requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
filelock==3.12.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python==4.7.0.72
Pillow==9.5.0
python-vlc==3.0.18122
sympy==1.12
torch==2.0.1
typing_extensions==4.6.3
